requirements.txt
# Requirements para Análisis Exploratorio de Datos y Métricas Avanzadas
# Sistema Recomendador Multimodal - H&M Fashion Recommendations

# Data manipulation and analysis
pandas>=1.5.0
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Scientific computing and statistics
scipy>=1.9.0
scikit-learn>=1.1.0

# Advanced metrics and network analysis
networkx>=2.8.0  # for graph theory and network analysis
powerlaw>=1.4.6  # for power-law distribution analysis

# Additional utilities
pathlib2>=2.3.0

# Image processing and utilities for normalization
Pillow>=9.0.0
tqdm>=4.64.0

# Optional: for advanced visualizations
kaleido>=0.2.1  # for plotly static image export